Move the TestServer header
