+++ tests/miral/active_window.cpp
@@ -16,7 +16,7 @@
  * Authored by: Alan Griffiths <alan@octopull.co.uk>
  */
 
-#include "test_server.h"
+#include <miral/test_server.h>
 
 #include <mir/client/connection.h>
 #include <mir/client/surface.h>
