+++ tests/miral/test_server.cpp
@@ -16,7 +16,7 @@
  * Authored by: Alan Griffiths <alan@octopull.co.uk>
  */
 
-#include "test_server.h"
+#include <miral/test_server.h>
 #include "basic_window_manager.h"
 #include "window_management_trace.h"
